deployment/pyproject.toml
[tool.poetry]
name = "deployment"
version = "0.1.0"
description = ""
authors = ["Rocky Jaiswal <rocky.jaiswal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"aws-cdk.core" = "^1.116.0"
"aws-cdk.aws-lambda" = "^1.116.0"
"aws-cdk.aws-apigateway" = "^1.116.0"
"aws-cdk.aws-apigatewayv2" = "^1.116.0"
"aws-cdk.aws-apigatewayv2-integrations" = "^1.116.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
